+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -13,14 +13,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.dto.ReservationTimeUserWebResponse;
+import roomescape.controller.dto.ReservationTimeWebRequest;
+import roomescape.controller.dto.ReservationTimeWebResponse;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.app.ReservationTimeAppRequest;
-import roomescape.dto.app.ReservationTimeAppResponse;
-import roomescape.dto.web.ReservationTimeUserWebResponse;
-import roomescape.dto.web.ReservationTimeWebRequest;
-import roomescape.dto.web.ReservationTimeWebResponse;
 import roomescape.exception.ReservationExistsException;
 import roomescape.service.ReservationTimeService;
+import roomescape.service.dto.ReservationTimeAppRequest;
+import roomescape.service.dto.ReservationTimeAppResponse;
 
 @RestController
 @RequestMapping("/times")
