+++ src/main/java/roomescape/controller/dto/ReservationTimeWebRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.web;
+package roomescape.controller.dto;
 
 public record ReservationTimeWebRequest(Long id, String startAt) {
