+++ src/main/java/roomescape/controller/dto/ReservationTimeWebResponse.java
@@ -1,13 +1,13 @@
-package roomescape.dto.web;
+package roomescape.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 import roomescape.domain.Reservation;
 
 public record ReservationTimeWebResponse(
-        Long id,
-        @JsonFormat(pattern = "HH:mm")
-        LocalTime startAt) {
+    Long id,
+    @JsonFormat(pattern = "HH:mm")
+    LocalTime startAt) {
 
     public static ReservationTimeWebResponse from(Reservation reservation) {
         return new ReservationTimeWebResponse(reservation.getTime().getId(), reservation.getTime().getStartAt());
