+++ src/main/java/roomescape/controller/dto/ReservationWebRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.web;
+package roomescape.controller.dto;
 
 public record ReservationWebRequest(String name, String date, Long timeId, Long themeId) {
