+++ src/main/java/roomescape/controller/dto/ThemeWebRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.web;
+package roomescape.controller.dto;
 
 public record ThemeWebRequest(String name, String description, String thumbnail) {
