+++ src/main/java/roomescape/service/dto/ReservationAppRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.app;
+package roomescape.service.dto;
 
 public record ReservationAppRequest(String name, String date, Long timeId, Long themeId) {
