+++ src/main/java/roomescape/service/dto/ReservationTimeAppRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.app;
+package roomescape.service.dto;
 
 public record ReservationTimeAppRequest(String startAt) {
