+++ src/main/java/roomescape/service/dto/ReservationTimeAppResponse.java
@@ -1,11 +1,12 @@
-package roomescape.dto.web;
+package roomescape.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 
-public record ReservationTimeUserWebResponse(
+public record ReservationTimeAppResponse(
     Long id,
     @JsonFormat(pattern = "HH:mm") LocalTime startAt,
-    boolean alreadyBooked) {
+    boolean alreadyBooked
+) {
 
 }
