+++ src/main/java/roomescape/service/dto/ThemeAppRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.app;
+package roomescape.service.dto;
 
 public record ThemeAppRequest(String name, String description, String thumbnail) {
